Fix off-by-one padding bug
If `add_padding()` is allowed to return a path with a trailing path
separator, it will get collapsed elsewhere in Spack. This can lead to
buildcache entries that have RPATHS that are too short to be replaced by
other users whose install root happens to be padded to the correct
length.  Detect this and replace the trailing path separator with an
concrete path character.

Signed-off-by: Samuel E. Browne <sebrown@sandia.gov>